requirements.txt
langchain==0.0.228 
openai==0.27.8 
faiss-cpu==1.7.4
tiktoken==0.4.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.25.1
pytesseract==0.3.10
streamlit==1.24.1
streamlit-toggle-switch==1.0.2
ipython==7.12.0
pypdf==3.15.0